requirements.txt
opencv-python-headless 
pillow